# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "description": "Api Documentation",
        "contact": {
            "name": "xyz",
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/search_addmission_no": {
            "post": {
                "tags": [
                    "Student"
                ],
                "summary": "Search Student",
                "operationId": "searchStdByAdmNo",
                "parameters": [
                    {
                        "name": "admissionNo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}